+++ backend/src/main/java/com/bluepa/backend/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.bluepa.backend.user.domain;
+package com.bluepa.backend.user.repository;
 
+import com.bluepa.backend.user.domain.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
